requirements-RTX5060.txt
numpy==1.26.4
soundfile==0.12.1
faster-whisper==0.10.0
onnxruntime-gpu==1.16.3
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2